Profile Page Updated. Buds fixed. Useless files deleted.
