--- VK-App/UI-Home Application/FriendsSectionHeader.swift
@@ -1,18 +1,0 @@
-//
-//  FriendsSectionHeader.swift
-//  UI-Home Application
-//
-//  Created by Developer on 22/03/2019.
-//  Copyright © 2019 Developer. All rights reserved.
-////
-//
-//import UIKit
-//
-//class FriendsSectionHeader: UITableViewHeaderFooterView {
-//    var label : UILabel = UILabel(frame: CGRect(x: 0, y: 0, width: 100, height: 20))
-//
-//    override func awakeFromNib() {
-//        super.awakeFromNib()
-//        self.addSubview(label)
-//    }
-//}
